problem: with build on Linux (NIXOS)

+++ lib/widgets/app_drawer.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:provider/provider.dart';
 import 'package:simpledictionary/l10n/app_localizations.dart';
-import 'package:simpledictionary/theme/app_theme.dart';
 
 import '../providers/dictionary_provider.dart';
 import '../screens/about_app_screen.dart';
@@ -31,7 +30,7 @@ class _AppDrawerState extends State<AppDrawer> {
             decoration: BoxDecoration(
               color: Theme.of(
                 context,
-              ).colorScheme.primary.withValues(alpha: 0.1.toInt()),
+              ).colorScheme.primary.withValues(alpha: 0.1),
             ),
             child: Row(
               crossAxisAlignment: CrossAxisAlignment.center,
@@ -144,7 +143,7 @@ class _DrawerInfoItem extends StatelessWidget {
             size: 24,
             color: Theme.of(
               context,
-            ).textTheme.bodySmall?.color?.withValues(alpha: 0.7.toInt()),
+            ).textTheme.bodySmall?.color?.withValues(alpha: 0.7),
           ),
           const SizedBox(width: 16),
           Text(text, style: Theme.of(context).textTheme.bodyLarge),
